Issue #18104: Remove old syntax from OperatorWrapCheck
